Properly abstract the ManaHandler structure
Really, only the implementations that are intended to directly handle
mana as capabilities need access to the updateClient method, right?

Also abstracts out handling NBT, because I already have a good system in
place for making things with the same data structure as a handler, so
why not use it?